requirements.txt
apache-airflow==2.7.1
requests==2.31.0
psycopg2-binary==2.9.9
pandas==2.1.1
python-dotenv==1.0.0
